+++ client/src/utils/auth.js
@@ -1,0 +1,2 @@
+// utils/auth.js
+export const isAuthenticated = () => !!localStorage.getItem('token');
